requirements.txt
fastapi==[0.116.1]
uvicorn==[0.35.0]
psycopg2-binary==[2.9.10]
deepface==[0.0.95]
pydantic==[2.11.7]